requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain>=0.1.0
google-generativeai
PyPDF2
faiss-cpu
appwrite
